+++ packages/flutter_robusta_hive/test/src/extension_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter_robusta_hive/flutter_robusta_hive.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
@@ -20,31 +22,42 @@ void main() {
 
       final runner = Runner(
         extensions: [
-          EventExtension<RunEvent>({
-            (e) => expectLater(() => Hive.openBox('test'), returnsNormally): 0,
-          }),
-          EventExtension<ExceptionEvent>({
-            (e) => hasError = true: 0,
-          }),
+          EventExtension(
+            configurator: (em, c) {
+              em.addEventListener<RunEvent>(
+                (e) => Hive.openBox('test'),
+              );
+              em.addEventListener<ExceptionEvent>(
+                (e) => hasError = true,
+              );
+            },
+          )
         ],
       );
 
-      await runner.run();
+      unawaited(runner.run());
 
-      expect(hasError, isTrue);
+      await expectLater(
+        Future.delayed(Duration.zero, () => hasError),
+        completion(isTrue),
+      );
     });
 
     test('can open box when have extension', () async {
       final runner = Runner(
         extensions: [
           FlutterHiveExtension(),
-          EventExtension<RunEvent>({
-            (e) => expectLater(() => Hive.openBox('test'), returnsNormally): 0,
-          }),
+          EventExtension(
+            configurator: (em, c) {
+              em.addEventListener<RunEvent>(
+                (e) => expectLater(Hive.openBox('test'), completes),
+              );
+            },
+          )
         ],
       );
 
-      await runner.run();
+      await expectLater(runner.run(), completes);
     });
   });
 }