+++ packages/robusta_runner/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0
 - Add dependence extension.
+- BC break: normalize `EventExtension` and `ImplementingCallbackExtension`.
 
 ## 0.1.1
 - Sealed base classes.